+++ lib/src/components/tf_checkbox_group.dart
@@ -10,6 +10,7 @@ class TFCheckboxGroup<T> extends StatefulWidget {
   final void Function(List<T>) onChanged;
   final List<TFValidationType> validationTypes;
   final TextEditingController? relatedController;
+  final TFGroupStyle? style;
 
   TFCheckboxGroup({
     Key? key,
@@ -18,7 +19,8 @@ class TFCheckboxGroup<T> extends StatefulWidget {
     required this.onChanged,
     this.initialValues,
     this.validationTypes = const <TFValidationType>[],
-    this.relatedController,
+    this.relatedController,  
+    this.style,
   }) : super(key: key) {
     if (validationTypes.contains(TFValidationType.requiredIfHas) &&
         relatedController == null) {
@@ -108,7 +110,7 @@ class _TFCheckboxGroupState<T> extends State<TFCheckboxGroup<T>> {
       children: [
         Text(
           widget.title,
-          style: TFFormStyle.of(context).titleStyle,
+          style: _tffStyle.titleStyle,
         ),
         const SizedBox(height: 8),
         ...List.generate(widget.items.length, (index) {
@@ -126,13 +128,13 @@ class _TFCheckboxGroupState<T> extends State<TFCheckboxGroup<T>> {
     return CheckboxListTile(
       title: Text(
         item.title,
-        style: TFFormStyle.of(context).groupStyle.itemTitleStyle.copyWith(
-              color: _isValid ? null : TFFormStyle.of(context).errorColor,
-            ),
+        style: _itemTitleStyle.copyWith(
+          color: _isValid ? null : _tffStyle.errorColor,
+        ),
       ),
       value: _selectedValues.contains(item.value),
       controlAffinity: ListTileControlAffinity.leading,
-      activeColor: TFFormStyle.of(context).activeColor,
+      activeColor: _tffStyle.activeColor,
       contentPadding: EdgeInsets.zero,
       checkboxShape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(2),
@@ -140,12 +142,15 @@ class _TFCheckboxGroupState<T> extends State<TFCheckboxGroup<T>> {
       side: BorderSide(
         width: 1.5,
         color: _isValid
-            ? TFFormStyle.of(context).groupStyle.unselectedColor
-            : TFFormStyle.of(context).errorColor,
+            ? _unselectedColor
+            : _tffStyle.errorColor,
       ),
       onChanged: (bool? isSelected) {
         _onItemChanged(item.value, isSelected ?? false);
       },
     );
   }
+
+  get _itemTitleStyle => widget.style?.itemTitleStyle ?? _tffStyle.groupStyle.itemTitleStyle;
+  get _unselectedColor => widget.style?.unselectedColor ?? _tffStyle.groupStyle.unselectedColor;
 }